Cleaned up codes.
